move localization exports to index
