+++ crowdin/messages.json
@@ -1,1 +1,1 @@
-{"262016571":"secondary button example","335040248":"About us","494177212":"You just hit a route that doesn","1622406547":"404: Not found","1808393236":"Login","-442695579":"Footer","-1076138910":"Trade","-1642457320":"Help centre","-1322632653":"NOT FOUND","-773544978":"Home"}
+{"262016571":"secondary button example","335040248":"About us","1622406547":"404: Not found","1808393236":"Login","-442695579":"Footer","-1076138910":"Trade","-1642457320":"Help centre","-1322632653":"NOT FOUND","-773544978":"Home"}
