+++ gatsby-browser.js
@@ -1,4 +1,4 @@
-import { WrapPagesWithLocaleContext } from './src/components/localization/locale-context'
+import { WrapPagesWithLocaleContext } from './src/components/localization'
 import { BinarySocketGeneral } from './src/common/websocket/socket_general'
 import { NetworkMonitorBase } from './src/common/websocket/network_base'
 import './src/components/localization/config'
