+++ gatsby-ssr.js
@@ -1,4 +1,4 @@
-import { WrapPagesWithLocaleContext } from './src/components/localization/locale-context'
+import { WrapPagesWithLocaleContext } from './src/components/localization'
 import './src/components/localization/config'
 
 export const wrapPageElement = WrapPagesWithLocaleContext
