+++ src/components/localization/index.js
@@ -1,0 +1,5 @@
+export * from './locale-context'
+export * from './language-switcher'
+export * from './localized-link'
+export * from './config'
+export * from './withIntl'
