+++ src/components/localization/locale-context.js
@@ -1,9 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import GlobalStyle from '../../themes/global-style'
-const LocaleContext = React.createContext()
 
-const LocaleContextWrapper = ({ children, pageContext: { locale } }) => (
+export const LocaleContext = React.createContext()
+
+export const LocaleContextWrapper = ({ children, pageContext: { locale } }) => (
     <LocaleContext.Provider value={{ locale }}>
         <GlobalStyle />
         {children}
@@ -11,8 +12,6 @@ const LocaleContextWrapper = ({ children, pageContext: { locale } }) => (
 )
 
 // Language passed here from createPages
-const WrapPagesWithLocaleContext = ({ element, props }) => {
+export const WrapPagesWithLocaleContext = ({ element, props }) => {
     return <LocaleContextWrapper {...props}>{element}</LocaleContextWrapper>
 }
-
-export { LocaleContext, LocaleContextWrapper, WrapPagesWithLocaleContext }
